+++ src/app/(detail)/profile/page.tsx
